// BassLines.Web/src/data/src/models/SpotifyClientAuth.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BassLines
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SpotifyClientInitialAuth
 */
export interface SpotifyClientAuth {
    /**
     * 
     * @type {string}
     * @memberof SpotifyClientAuth
     */
    accessToken?: string | null;
    /**
     * 
     * @type {number}
     * @memberof SpotifyClientAuth
     */
    expiryTime?: number;
    /**
     * 
     * @type {string}
     * @memberof SpotifyClientAuth
     */
    refreshToken?: string | null;
}

export function SpotifyClientAuthFromJSON(json: any): SpotifyClientAuth {
    return SpotifyClientAuthFromJSONTyped(json, false);
}

export function SpotifyClientAuthFromJSONTyped(json: any, ignoreDiscriminator: boolean): SpotifyClientAuth {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'accessToken': !exists(json, 'accessToken') ? undefined : json['accessToken'],
        'expiryTime': !exists(json, 'expiryTime') ? undefined : json['expiryTime'],
        'refreshToken': !exists(json, 'refreshToken') ? undefined : json['refreshToken'],
    };
}

export function SpotifyClientAuthToJSON(value?: SpotifyClientAuth | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'accessToken': value.accessToken,
        'expiryTime': value.expiryTime,
        'refreshToken': value.refreshToken,
    };
}